43-collocation-of-project-structure-2 ・プロジェクトを技術駆動パッケージングからコロケーション化
