+++ lib/features/chat/presentation/providers/ai_service_provider.dart
@@ -1,0 +1,7 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+import 'package:alexa_to_ai/features/chat/data/services/ai_service.dart';
+
+final aiServiceProvider = Provider<AIService>((ref) {
+  return AIService();
+});
