+++ lib/features/settings/presentation/providers/biometric_authentication_service_provider.dart
@@ -1,0 +1,7 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+import 'package:alexa_to_ai/features/settings/data/services/biometric_authentication_service.dart';
+
+final biometricAuthProvider = Provider<BiometricAuthenticationService>((ref) {
+  return BiometricAuthenticationService();
+});
