+++ lib/main.dart
@@ -1,8 +1,8 @@
-import 'package:alexa_to_ai/database/database.dart';
-import 'package:alexa_to_ai/services/login_authentication_service.dart';
-import 'package:alexa_to_ai/widgets/navigation/footer.dart';
-import 'package:alexa_to_ai/widgets/theme/dark_theme_data.dart';
-import 'package:alexa_to_ai/widgets/theme/light_theme_data.dart';
+import 'package:alexa_to_ai/app/navigation/footer.dart';
+import 'package:alexa_to_ai/app/theme/dark_theme_data.dart';
+import 'package:alexa_to_ai/app/theme/light_theme_data.dart';
+import 'package:alexa_to_ai/features/authentication/data/services/login_authentication_service.dart';
+import 'package:alexa_to_ai/features/settings/data/local/settings_hive_box.dart';
 import 'package:amplify_authenticator/amplify_authenticator.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
@@ -17,12 +17,8 @@ void main() async {
 }
 
 Future<void> someAsyncFunction() async {
-  // .envファイルの読み込み
   await dotenv.load();
-  // ローカルデータベースを初期化
-  WidgetsFlutterBinding.ensureInitialized();
-  await initHive();
-  // Cognitoで認証を行うための事前処理
+  await initSettingsHiveBox();
   final loginAuthenticationService = LoginAuthenticationService();
   await loginAuthenticationService.configureAmplify();
 }
