--- lib/providers/biometric_authentication_service_provider.dart
@@ -1,7 +1,0 @@
-import 'package:alexa_to_ai/services/biometric_authentication_service.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-// シングルトンインスタンスを提供するProvider
-final biometricAuthProvider = Provider<BiometricAuthenticationService>((ref) {
-  return BiometricAuthenticationService();
-});
