--- lib/widgets/barrel/chat_ai_screen_widgets.dart
@@ -1,4 +1,0 @@
-export 'package:alexa_to_ai/widgets/notification/alert_dialog.dart';
-export 'package:alexa_to_ai/widgets/theme/custom_dark_chat_theme.dart';
-export 'package:alexa_to_ai/widgets/theme/custom_default_chat_theme.dart';
-export 'package:flutter/material.dart';
